+++ CHANGELOG.md
@@ -11,6 +11,7 @@
 - feat(cocoa): Update snippets to use new UI Profiling configureation ([#933](https://github.com/getsentry/sentry-wizard/pull/933))
 - fix(react-native): Handles xcode build phase patching failure ([#866](https://github.com/getsentry/sentry-wizard/pull/866))
 - feat(react-native): Add a Session Replay step ([#915](https://github.com/getsentry/sentry-wizard/pull/915))
+- feat(nextjs): Add connectivity check to example page([#XXX](https://github.com/getsentry/sentry-wizard/pull/XXX))
 
 ## 4.6.0